[009-FEAT] : Category

+++ src/main/java/com/example/board/service/impl/RankUpStandardImpl.java
@@ -5,15 +5,18 @@ import static com.example.board.exception.ErrorCode.ALREADY_RANK;
 import static com.example.board.exception.ErrorCode.NOT_ADD_RANK;
 import static com.example.board.exception.ErrorCode.NOT_FIND_RANK;
 import static com.example.board.exception.ErrorCode.OVER_RANK_CHECK_BOARD_COMMENT;
+import static com.example.board.exception.ErrorCode.REGISTERED_RANK_CATEGORY;
 import static com.example.board.exception.ErrorCode.UNDER_RANK_CHECK_BOARD_COMMENT;
 
-import com.example.board.domain.dto.RankUpStandardDto.SearchRankUpStandard;
+import com.example.board.domain.dto.RankUpStandardDto;
 import com.example.board.domain.entity.RankUpStandard;
 import com.example.board.domain.entity.User;
 import com.example.board.domain.form.RankUpStandardForm.AddRankUpStandard;
 import com.example.board.domain.form.RankUpStandardForm.ModifyRankUpStandard;
+import com.example.board.domain.repository.CategoryRepository;
 import com.example.board.domain.repository.RankUpStandardRepository;
 import com.example.board.domain.type.RankType;
+import com.example.board.exception.ErrorCode;
 import com.example.board.exception.GlobalException;
 import com.example.board.service.RankUpStandardService;
 import com.example.board.service.UserTypeCheckService;
@@ -30,15 +33,16 @@ public class RankUpStandardImpl implements RankUpStandardService {
 
     private final UserTypeCheckService userTypeCheckService;
     private final RankUpStandardRepository rankUpStandardRepository;
+    private final CategoryRepository categoryRepository;
 
 
     // 조회
-    public List<SearchRankUpStandard> searchRankUpStandard() {
+    public List<RankUpStandardDto> searchRankUpStandard() {
 
-        List<SearchRankUpStandard> infoRankUpStandards = new ArrayList<>();
+        List<RankUpStandardDto> infoRankUpStandards = new ArrayList<>();
 
         for (RankUpStandard rankUpStandard : rankUpStandardRepository.findAllByOrderByRankName()) {
-            SearchRankUpStandard searchRankUpStandard = SearchRankUpStandard.builder()
+            RankUpStandardDto searchRankUpStandard = RankUpStandardDto.builder()
                 .standardId(rankUpStandard.getStandardId())
                 .rankName(rankUpStandard.getRankName())
                 .boardCount(rankUpStandard.getBoardCount())
@@ -55,7 +59,7 @@ public class RankUpStandardImpl implements RankUpStandardService {
 
     // 저장
     @Transactional
-    public List<SearchRankUpStandard> addRankUpStandard(
+    public List<RankUpStandardDto> addRankUpStandard(
         Long userId, AddRankUpStandard addRankUpStandard) {
 
         if (rankUpStandardRepository.existsByRankName(addRankUpStandard.getRankName())) {
@@ -86,7 +90,7 @@ public class RankUpStandardImpl implements RankUpStandardService {
 
     // 수정
     @Transactional
-    public List<SearchRankUpStandard> modifyRankUpStandard(
+    public List<RankUpStandardDto> modifyRankUpStandard(
         Long userId, Long standardId, ModifyRankUpStandard modifyRankUpStandard) {
 
         User user = userTypeCheckService.userTypeAdmin(userId);
@@ -109,12 +113,16 @@ public class RankUpStandardImpl implements RankUpStandardService {
 
     // 삭제
     @Transactional
-    public List<SearchRankUpStandard> deleteRankUpStandard(Long userId, Long standardId) {
+    public List<RankUpStandardDto> deleteRankUpStandard(Long userId, Long standardId) {
 
         userTypeCheckService.userTypeAdmin(userId);
 
         RankUpStandard rankUpStandard = rankUpStandardRepository.findByStandardId(standardId)
             .orElseThrow(() -> new GlobalException(NOT_FIND_RANK));
+
+        if (categoryRepository.existsByCategoryRank(rankUpStandard.getRankName())) {
+            throw new GlobalException(REGISTERED_RANK_CATEGORY);
+        }
 
         rankUpStandardRepository.delete(rankUpStandard);
